Move specs from Germany to USA as default country
